remove unnecessary main module, moved it to app module
