+++ app/build.gradle.kts
@@ -68,7 +68,6 @@ dependencies {
     androidTestImplementation(libs.android.junit)
 
     // other modules
-    implementation(project(":feature-main"))
     implementation(project(":feature-cash-book"))
     implementation(project(":feature-statistics"))
 }
