+++ app/src/main/kotlin/com/farmer/olive/navigation/OliveNavHost.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.ComposeNavigator
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
-import com.farmer.feature_main.OliveScreens
+import com.farmer.olive.ui.OliveScreens
 import com.farmer.feature_statistics.StatisticsScreen
 import com.farmer.home.CashbookScreen
 
@@ -14,16 +14,12 @@ fun OliveNavHost(
     navHostController: NavHostController,
     composeNavigator: ComposeNavigator
 ) {
-    NavHost(navController = navHostController, startDestination = OliveScreens.Main.route) {
+    NavHost(navController = navHostController, startDestination = OliveScreens.CashBook.route) {
         composable(route = "cash_book") {
             CashbookScreen()
         }
         composable(route = "statistics") {
             StatisticsScreen()
         }
-        oliveHomeNavigation(
-            composeNavigator = composeNavigator,
-            navController = navHostController
-        )
     }
 }
