+++ app/src/main/kotlin/com/farmer/olive/ui/OliveBottomNavigation.kt
@@ -1,0 +1,45 @@
+package com.farmer.olive.ui
+
+import androidx.compose.material.BottomNavigation
+import androidx.compose.material.BottomNavigationItem
+import androidx.compose.material.Icon
+import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.Favorite
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.navigation.NavController
+import androidx.navigation.NavDestination.Companion.hierarchy
+import androidx.navigation.NavGraph.Companion.findStartDestination
+import androidx.navigation.compose.currentBackStackEntryAsState
+
+@Composable
+fun OliveBottomNavigation(
+    navController: NavController,
+    screenList: List<OliveScreens>
+) {
+    BottomNavigation {
+        val navBackStackEntry by navController.currentBackStackEntryAsState()
+        val currentDestination = navBackStackEntry?.destination
+        screenList.forEach { screen ->
+            BottomNavigationItem(
+                icon = {
+                    Icon(Icons.Filled.Favorite, contentDescription = null)
+                },
+                label = {
+                    Text(screen.screenName)
+                },
+                selected = currentDestination?.hierarchy?.any { it.route == screen.route } == true,
+                onClick = {
+                    navController.navigate(screen.route) {
+                        popUpTo(navController.graph.findStartDestination().id) {
+                            saveState = true
+                        }
+                        launchSingleTop = true
+                        restoreState = true
+                    }
+                }
+            )
+        }
+    }
+}
