+++ app/src/main/kotlin/com/farmer/olive/ui/OliveMain.kt
@@ -1,6 +1,8 @@
 package com.farmer.olive.ui
 
+import androidx.compose.material.Scaffold
 import androidx.compose.runtime.Composable
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.navigation.compose.ComposeNavigator
 import androidx.navigation.compose.rememberNavController
 import com.farmer.olive.navigation.OliveNavHost
@@ -8,10 +10,17 @@ import com.farmer.olive.ui.theme.OliveTheme
 
 @Composable
 fun OliveMain(
-    composeNavigator: ComposeNavigator
+    composeNavigator: ComposeNavigator,
+    viewModel: MainViewModel = hiltViewModel()
 ) {
     OliveTheme {
         val navHostController = rememberNavController()
-        OliveNavHost(navHostController = navHostController, composeNavigator = composeNavigator)
+        Scaffold(
+            bottomBar = {
+                OliveBottomNavigation(navHostController, viewModel.oliveScreens)
+            }
+        ) {
+            OliveNavHost(navHostController = navHostController, composeNavigator = composeNavigator)
+        }
     }
 }
