+++ app/src/main/kotlin/com/farmer/olive/ui/OliveScreens.kt
@@ -1,16 +1,10 @@
-package com.farmer.feature_main
+package com.farmer.olive.ui
 
 // todo screen name should be changed to string id
 sealed class OliveScreens(
     val route: String,
     val screenName: String = ""
 ) {
-    // home screen
-    object Home : OliveScreens("home")
-
-    // main screen
-    object Main : OliveScreens("main")
-
     // cash book screen
     object CashBook : OliveScreens("cash_book", "Cashbook")
