--- feature-main/build.gradle.kts
@@ -1,27 +1,0 @@
-plugins {
-    id("com.android.library")
-    id("org.jetbrains.kotlin.android")
-    kotlin("kapt")
-    id("dagger.hilt.android.plugin")
-}
-
-android {
-    namespace = "com.farmer.feature_main"
-}
-
-dependencies {
-
-    // base
-    implementation(libs.bundles.compose)
-    implementation(libs.core.ktx)
-    implementation(libs.lifecycle)
-    implementation(libs.lifecycle.viewmodel)
-
-    // hilt
-    implementation(libs.hilt.android)
-    kapt(libs.hilt.android)
-    kapt(libs.hilt.compiler)
-
-    testImplementation(libs.junit)
-    androidTestImplementation(libs.android.junit)
-}
