--- feature-main/src/main/kotlin/com/farmer/feature_main/MainScreen.kt
@@ -1,61 +1,0 @@
-package com.farmer.feature_main
-
-import androidx.compose.material.BottomNavigation
-import androidx.compose.material.BottomNavigationItem
-import androidx.compose.material.Icon
-import androidx.compose.material.Scaffold
-import androidx.compose.material.Text
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Favorite
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.navigation.NavController
-import androidx.navigation.NavDestination.Companion.hierarchy
-import androidx.navigation.NavGraph.Companion.findStartDestination
-import androidx.navigation.compose.ComposeNavigator
-import androidx.navigation.compose.currentBackStackEntryAsState
-
-@Composable
-fun MainScreen(
-    composeNavigator: ComposeNavigator,
-    navController: NavController,
-    viewModel: MainViewModel = hiltViewModel()
-) {
-    Scaffold(
-        bottomBar = {
-            BottomNavigation {
-                val navBackStackEntry by navController.currentBackStackEntryAsState()
-                val currentDestination = navBackStackEntry?.destination
-                viewModel.oliveScreens.forEach { screen ->
-                    BottomNavigationItem(
-                        icon = {
-                            Icon(Icons.Filled.Favorite, contentDescription = null)
-                        },
-                        label = {
-                            Text(screen.screenName)
-                        },
-                        selected = currentDestination?.hierarchy?.any { it.route == screen.route } == true,
-                        onClick = {
-                            navController.navigate(screen.route) {
-                                // Pop up to the start destination of the graph to
-                                // avoid building up a large stack of destinations
-                                // on the back stack as users select items
-                                popUpTo(navController.graph.findStartDestination().id) {
-                                    saveState = true
-                                }
-                                // Avoid multiple copies of the same destination when
-                                // reselecting the same item
-                                launchSingleTop = true
-                                // Restore state when reselecting a previously selected item
-                                restoreState = true
-                            }
-                        }
-                    )
-                }
-            }
-        }
-    ) {
-        Text("Hello Main Screen!")
-    }
-}
