+++ settings.gradle.kts
@@ -27,5 +27,4 @@ dependencyResolutionManagement {
 rootProject.name = "Olive"
 include(":app")
 include(":feature-cash-book")
-include(":feature-main")
 include(":feature-statistics")
